Simple command line application to exercise text extractors + refactoring
