+++ logicle/lib/chat/conversion.ts
@@ -5,15 +5,8 @@ import * as dto from '@/types/dto'
 import { logger } from '@/lib/logging'
 import { storage } from '@/lib/storage'
 import { LlmModelCapabilities } from './models'
-import mammoth from 'mammoth'
-import ExcelJS from 'exceljs'
-import { ensureABView } from '../utils'
-import { sheetToMarkdown } from '../xlstomarkdown'
-import micromatch from 'micromatch'
+import { findExtractor } from '../textextraction'
 import env from '../env'
-import { gfm } from 'turndown-plugin-gfm'
-import TurndownService, { Node as TurndownNode } from 'turndown'
-import { PDFExtract } from 'pdf.js-extract'
 
 const loadImagePartFromFileEntry = async (fileEntry: schema.File): Promise<ai.ImagePart> => {
   const fileContent = await storage.readBuffer(fileEntry.path, !!fileEntry.encrypted)
@@ -32,116 +25,6 @@ const loadFilePartFromFileEntry = async (fileEntry: schema.File): Promise<ai.Fil
     mediaType: fileEntry.type,
   }
   return image
-}
-
-type Converter = (buffer: Buffer) => Promise<string>
-
-function hasGetAttribute(n: unknown): n is { getAttribute(name: string): string | null } {
-  return !!n && typeof (n as any).getAttribute === 'function'
-}
-
-function createTurndown() {
-  const td = new TurndownService({
-    headingStyle: 'atx', // # H1, ## H2, ...
-    codeBlockStyle: 'fenced', // ``` fenced blocks
-    bulletListMarker: '-', // -, *, or +
-    emDelimiter: '_', // _emphasis_ vs *emphasis*
-    br: '\n', // <br> → newline
-  })
-
-  // Enable GitHub-flavored Markdown features
-  td.use([gfm])
-
-  td.addRule('fencedCodeWithLanguage', {
-    filter: (node) => {
-      return node.nodeName === 'PRE' && (node as HTMLElement).firstElementChild?.nodeName === 'CODE'
-    },
-    replacement: (_content, node) => {
-      const codeEl = (node as HTMLElement).firstElementChild as HTMLElement
-      const className = (codeEl.getAttribute('class') || '').toLowerCase()
-
-      // Extract language (supports language-xxx or lang-xxx)
-      const lang =
-        className.match(/language-([\w-]+)/)?.[1] ||
-        className.match(/lang(?:uage)?-([\w-]+)/)?.[1] ||
-        ''
-
-      const codeText = (codeEl.textContent || '').trimEnd()
-      const fence = '```'
-
-      return `\n${fence}${lang}\n${codeText}\n${fence}\n`
-    },
-  })
-  td.addRule('cleanImages', {
-    filter: 'img',
-    replacement: (_content: string, node: TurndownNode) => {
-      // Narrow the turndown node to something with getAttribute
-      if (!hasGetAttribute(node)) return '' // or fall back to default behavior
-
-      const alt = (node.getAttribute('alt') || '').replace(/\n/g, ' ')
-      const src = node.getAttribute('src') || ''
-      const title = node.getAttribute('title')
-
-      return title ? `![${alt}](${src} "${title}")` : `![${alt}](${src})`
-    },
-  }) // Optionally keep line breaks inside paragraphs (helps with Word line breaks)
-  td.keep(['br'])
-
-  return td
-}
-
-const wordConverter: Converter = async (data: Buffer) => {
-  const { value: html } = await mammoth.convertToHtml({
-    buffer: data,
-  })
-  const turndown = createTurndown()
-  const markdown = turndown.turndown(html)
-  return markdown
-}
-
-const excelConverter: Converter = async (data: Buffer) => {
-  const wb = new ExcelJS.Workbook()
-  await wb.xlsx.load(ensureABView(data).buffer)
-  const ws = wb.worksheets[0]
-  return sheetToMarkdown(ws, { headerRow: 1 })
-}
-
-const genericTextConverter: Converter = async (data: Buffer) => {
-  return data.toString('utf8')
-}
-
-const pdfConverter: Converter = async (data: Buffer) => {
-  return new Promise((resolve, reject) => {
-    const pdfExtract = new PDFExtract()
-    const options = { normalizeWhitespace: true }
-
-    pdfExtract.extractBuffer(data, options, (err, data) => {
-      if (err) return reject(err)
-      if (data) {
-        const text = data.pages
-          .map((page) => page.content.map((item) => item.str).join(' '))
-          .join('\n\n')
-
-        resolve(text)
-      }
-    })
-  })
-}
-
-export const converters: Record<string, Converter> = {
-  ['application/vnd.openxmlformats-officedocument.wordprocessingml.document']: wordConverter,
-  ['application/vnd.openxmlformats-officedocument.spreadsheetml.sheet']: excelConverter,
-  ['application/pdf']: pdfConverter,
-  ['text/*']: genericTextConverter,
-}
-
-const findConverter = (desiredFileType: string) => {
-  if (!env.chat.enableAttachmentConversion) {
-    return undefined
-  }
-  return Object.entries(converters).find(([fileType, _converter]) => {
-    return micromatch.isMatch(desiredFileType, fileType)
-  })?.[1]
 }
 
 // Not easy to do it right... Claude will crash if the input image format is not supported
@@ -233,15 +116,17 @@ export const dtoMessageToLlmMessage = async (
           if (capabilities.supportedMedia?.includes(fileEntry.type)) {
             return loadFilePartFromFileEntry(fileEntry)
           } else {
-            const converter = findConverter(fileEntry.type)
-            if (converter) {
-              const fileContent = await storage.readBuffer(fileEntry.path, !!fileEntry.encrypted)
-              const text = await converter(fileContent)
-              if (text) {
-                return {
-                  type: 'text',
-                  text: `Here is the text content of the file "${fileEntry.name}" with id ${fileEntry.id}\n${text}`,
-                } satisfies ai.TextPart
+            if (env.chat.enableAttachmentConversion) {
+              const converter = findExtractor(fileEntry.type)
+              if (converter) {
+                const fileContent = await storage.readBuffer(fileEntry.path, !!fileEntry.encrypted)
+                const text = await converter(fileContent)
+                if (text) {
+                  return {
+                    type: 'text',
+                    text: `Here is the text content of the file "${fileEntry.name}" with id ${fileEntry.id}\n${text}`,
+                  } satisfies ai.TextPart
+                }
               }
             }
             return {
